Academic-Agent-model/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
pypdf
transformers
torch
numpy
langchain-text-splitters
openai
python-dotenv
